fix: change edge weight type from int to float in Graph interface

+++ src/graph.h
@@ -26,10 +26,10 @@ public:
   virtual void print_graph() = 0;
   virtual vector<int> get_neighbors(int vertex) = 0;
 
-  virtual bool add_edge(int from_vertex, int to_vertex, int weight) = 0;
+  virtual bool add_edge(int from_vertex, int to_vertex, float weight) = 0;
   virtual bool remove_edge(int from_vertex, int to_vertex) = 0;
   virtual bool has_edge(int from_vertex, int to_vertex) = 0;
-  virtual int get_edge_weight(int from_vertex, int to_vertex) = 0;
+  virtual float get_edge_weight(int from_vertex, int to_vertex) = 0;
 
   string get_vertex_label(int vertex)
   {
